the FUEL horde
all this really is is just the fuel items; i am still yet to get the textures. it's been easier to do on the web because i can just make it do a HTTP request to github and it'll give me the texture.

i also have to think about portability; maybe i'll use rres, or a zip (using miniz)